refactor: simplify and consolidate handler logic
