+++ handlr-regex/Cargo.toml
@@ -29,6 +29,7 @@ regex = { version = "1", default-features = false, features = ["std", "unicode-p
 tabled = "0.15.0"
 serde_json = "1.0"
 serde_regex = "1.1.0"
+enum_dispatch = "0.3.13"
 
 [[bin]]
 name = "handlr"
