+++ handlr-regex/src/common/mod.rs
@@ -7,7 +7,9 @@ mod table;
 
 pub use self::db::autocomplete as db_autocomplete;
 pub use desktop_entry::{DesktopEntry, Mode as ExecMode};
-pub use handler::{DesktopHandler, Handler};
+pub use handler::{
+    DesktopHandler, Handleable, Handler, RegexApps, RegexHandler,
+};
 pub use mime_types::{MimeOrExtension, MimeType};
 pub use path::{mime_table, UserPath};
 pub use table::render_table;
